Minor changes in MapBuilder. Began work on Morphed Textured Plane, where images don’t have to appear rectangular.
Scaling.gcx is being used to adjust MorphedTexturedPlane (henceforth referred to as MTPlane). n is the width. m is the height. I’m trying to find a function, that would allow me to transform a point in the trapezoid into a point in the rectangle, so as to display morphed rectangles. This is needed for how I’m planning to make ghosts look like.